Fixed right half hover not affecting left side
I used the has:() pseudo-class which lets me select elements that appear before in the DOM structure.
Basically, without has:() I would not be able to easily affect things in the left-half because in the html code it was written before the right-half. has:() is an experimental pseudo-class that lets me do this!
This is an alternative to .main-container .right-half:hover ~ .left-half #left-half-pc which looks like it should work, but does not!